+++ src/os2datascanner/projects/report/organizations/admin.py
@@ -28,8 +28,8 @@ class ReadOnlyAdminMixin:
 @admin.register(Alias)
 class AliasAdmin(ReadOnlyAdminMixin, admin.ModelAdmin):
     list_filter = ["_alias_type"]
-    list_display = ('user', 'account', '_alias_type', '_value')
-    readonly_fields = ('user', 'account', '_alias_type', '_value')
+    list_display = ('user', 'account', '_alias_type', '_value', 'shared')
+    readonly_fields = ('user', 'account', '_alias_type', '_value', 'shared')
 
 
 @admin.register(Account)